modulize env + torii cli world logic
