+++ crates/dojo-world/Cargo.toml
@@ -17,6 +17,7 @@ convert_case.workspace = true
 dojo-types = { path = "../dojo-types" }
 futures = "0.3.28"
 reqwest = { version = "0.11.18", default-features = false, features = [ "rustls-tls" ] }
+scarb.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 serde_with.workspace = true
