+++ crates/dojo-world/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod environment;
 pub mod manifest;
 pub mod migration;
 pub mod utils;
