+++ crates/sozo/src/commands/migrate.rs
@@ -1,9 +1,9 @@
 use anyhow::Result;
 use clap::Args;
+use dojo_world::environment::dojo_metadata_from_workspace;
 use scarb::core::Config;
 
 use super::options::account::AccountOptions;
-use super::options::dojo_metadata_from_workspace;
 use super::options::starknet::StarknetOptions;
 use super::options::world::WorldOptions;
 use crate::ops::migration;
