+++ crates/sozo/src/commands/options/mod.rs
@@ -1,79 +1,11 @@
-use scarb::core::{ManifestMetadata, Workspace};
-use serde::Deserialize;
-
 pub mod account;
 pub mod starknet;
 pub mod world;
 
-pub(crate) fn dojo_metadata_from_workspace(ws: &Workspace<'_>) -> Option<DojoMetadata> {
-    Some(ws.current_package().ok()?.manifest.metadata.dojo())
-}
-
-#[derive(Default, Deserialize, Debug, Clone)]
-pub(crate) struct DojoMetadata {
-    env: Option<Environment>,
-}
-
-#[derive(Default, Deserialize, Clone, Debug)]
-pub struct Environment {
-    rpc_url: Option<String>,
-    account_address: Option<String>,
-    private_key: Option<String>,
-    keystore_path: Option<String>,
-    keystore_password: Option<String>,
-    world_address: Option<String>,
-}
-
-impl Environment {
-    pub fn world_address(&self) -> Option<&str> {
-        self.world_address.as_deref()
-    }
-
-    pub fn rpc_url(&self) -> Option<&str> {
-        self.rpc_url.as_deref()
-    }
-
-    pub fn account_address(&self) -> Option<&str> {
-        self.account_address.as_deref()
-    }
-
-    pub fn private_key(&self) -> Option<&str> {
-        self.private_key.as_deref()
-    }
-
-    #[allow(dead_code)]
-    pub fn keystore_path(&self) -> Option<&str> {
-        self.keystore_path.as_deref()
-    }
-
-    pub fn keystore_password(&self) -> Option<&str> {
-        self.keystore_password.as_deref()
-    }
-}
-
-impl DojoMetadata {
-    pub fn env(&self) -> Option<&Environment> {
-        self.env.as_ref()
-    }
-}
-trait MetadataExt {
-    fn dojo(&self) -> DojoMetadata;
-}
-
-impl MetadataExt for ManifestMetadata {
-    fn dojo(&self) -> DojoMetadata {
-        self.tool_metadata
-            .as_ref()
-            .and_then(|e| e.get("dojo"))
-            .cloned()
-            .map(|v| v.try_into::<DojoMetadata>().unwrap_or_default())
-            .unwrap_or_default()
-    }
-}
-
 #[cfg(test)]
 mod test {
-    use super::*;
+    use dojo_world::environment::DojoMetadata;
+
     #[test]
     fn check_deserialization() {
         let metadata: DojoMetadata = toml::from_str(
