+++ crates/sozo/src/commands/system.rs
@@ -1,8 +1,8 @@
 use anyhow::Result;
 use clap::{Args, Subcommand};
+use dojo_world::environment::dojo_metadata_from_workspace;
 use scarb::core::Config;
 
-use super::options::dojo_metadata_from_workspace;
 use super::options::starknet::StarknetOptions;
 use super::options::world::WorldOptions;
 use crate::ops::system;
