+++ crates/sozo/src/ops/component.rs
@@ -1,9 +1,9 @@
 use anyhow::Result;
 use dojo_client::contract::world::WorldContractReader;
+use dojo_world::environment::Environment;
 use starknet::core::types::{BlockId, BlockTag};
 
 use crate::commands::component::ComponentCommands;
-use crate::commands::options::Environment;
 
 pub async fn execute(command: ComponentCommands, env_metadata: Option<Environment>) -> Result<()> {
     match command {
