+++ crates/sozo/src/ops/execute.rs
@@ -1,8 +1,8 @@
 use anyhow::{Context, Result};
 use dojo_client::contract::world::WorldContract;
+use dojo_world::environment::Environment;
 
 use crate::commands::execute::ExecuteArgs;
-use crate::commands::options::Environment;
 
 pub async fn execute(args: ExecuteArgs, env_metadata: Option<Environment>) -> Result<()> {
     let ExecuteArgs { system, calldata, world, starknet, account } = args;
